Added compatibility for non-double Scalar type
- Added FADBAD++ headers: does not work as expected

git-svn-id: svn+ssh://jrlserver.muse.aist.go.jp/home/svn/Software/SCD@72 5988d2c4-2d00-3947-b3d2-f5ac3ac2f174

+++ include/SCD/Matrix/SmallMatrix3x3Mod.h
@@ -163,12 +163,12 @@ namespace MAL_Default
 		explicit Matrix3x3Mod<T> (const T &roll, const T &pitch, const T &yaw)
         {
 
-            T cy(cos(double(yaw))); 
-            T sy(sin(double(yaw))); 
-            T cp(cos(double(pitch))); 
-            T sp(sin(double(pitch))); 
-            T cr(cos(double(roll)));
-            T sr(sin(double(roll)));
+            T cy(cos(yaw)); 
+            T sy(sin(yaw)); 
+            T cp(cos(pitch)); 
+            T sp(sin(pitch)); 
+            T cr(cos(roll));
+            T sr(sin(roll));
 
             T cc = cy * cr; 
             T cs = cy * sr;
